+++ config/models-scanner.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 return [
 
     'models_namespace_regex' => '/\\\Models\\\/i',