+++ routes/web.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 use Axn\ModelsScanner\Controllers\ScanController;
 use Illuminate\Support\Facades\Route;